+++ src/Control/Lens/Fold.hs
@@ -81,9 +81,10 @@ module Control.Lens.Fold
 
 import Control.Applicative as Applicative
 import Control.Applicative.Backwards
+import Control.Lens.Classes
 import Control.Lens.Getter
 import Control.Lens.Internal
-import Control.Lens.Unsafe
+import Control.Lens.Internal.Combinators
 import Control.Lens.Type
 import Control.Monad
 import Data.Foldable as Foldable