+++ src/Data/List/Split/Lens.hs
@@ -38,7 +38,7 @@ module Data.List.Split.Lens
 
 import Control.Applicative
 import Control.Lens
-import Control.Lens.Internal
+import Control.Lens.Classes
 import Data.Monoid
 import Data.List.Split
 import Data.List.Split.Internals